requirements.txt
flask
wtforms
werkzeug.utils
requests
unittest